Implemented a service for calculating the price including coupon and tax - Routing configuration, Symfony Validator, Tax Calculator

+++ src/Entity/Product.php
@@ -35,7 +35,7 @@ class /**/Product
         return $this;
     }
 
-    public function getPrice(): float
+    public function getPrice(): int
     {
         return $this->price;
     }
